Rename Processing UI state to InProgress
